backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dj-rest-auth==5.0.1
djangorestframework-simplejwt==5.3.1
django-filter==23.5
drf-yasg==1.21.7
Pillow==10.1.0
reportlab==4.0.9

